Options Menu Fixed and More Collision in Home.tscn
Please please please let this be the last fix the options menu needs.

I also deleted a huge img file of the house I was using earlier.

I got a bit side tracked and added a small easter egg (this can be removed lol)

Added the rest of the collision to the house placeholder